Create script.js
First release of this script.js.
The code should change the ext of the button in the html index file, once the button is pressed from
"Press this button" to "You pressed the button".
The script ascribes an Id to the button. 
Pending now in the html index file: For this to work, I need to insert this id in the html index file.